tests/requirements_comprehensive.txt
# Comprehensive Test Suite Requirements for Echo Brain API Testing

# Core testing framework
pytest>=7.4.0
pytest-asyncio>=0.21.1
pytest-xdist>=3.3.1          # Parallel test execution
pytest-cov>=4.1.0            # Coverage reporting
pytest-html>=3.2.0           # HTML test reports
pytest-json-report>=1.5.0    # JSON test reports

# HTTP client for API testing
httpx>=0.24.1
requests>=2.31.0

# Database testing
asyncpg>=0.28.0
psycopg2-binary>=2.9.7

# Redis testing
redis>=4.6.0

# Performance monitoring
psutil>=5.9.5
memory-profiler>=0.60.0

# Security testing
safety>=2.3.1                # Dependency vulnerability scanning
bandit>=1.7.5                # Security linting

# JSON Web Tokens
PyJWT>=2.8.0

# Data manipulation and analysis
pandas>=2.0.3
numpy>=1.24.3

# Mocking and test utilities
responses>=0.23.3             # HTTP request mocking
freezegun>=1.2.2             # Time mocking
factory-boy>=3.3.0           # Test data generation

# Load testing
locust>=2.15.1               # Load testing framework

# Code quality
flake8>=6.0.0                # Linting
black>=23.7.0                # Code formatting
isort>=5.12.0                # Import sorting

# Documentation
sphinx>=7.1.2               # Documentation generation
sphinx-rtd-theme>=1.3.0     # Read the Docs theme

# Environment management
python-dotenv>=1.0.0

# Development utilities
ipython>=8.14.0
jupyter>=1.0.0

# Additional test utilities
colorama>=0.4.6              # Colored terminal output
tqdm>=4.65.0                 # Progress bars
click>=8.1.6                 # CLI framework